Replace old template with new one
