+++ packages/cna-template/template/.eslintrc.json
