--- packages/cna-template/template/babel.config.json
@@ -1,14 +1,0 @@
-{
-  "presets": [
-    [
-      "next/babel",
-      {
-        "preset-react": {
-          "runtime": "automatic",
-          "importSource": "@emotion/react"
-        }
-      }
-    ]
-  ],
-  "plugins": ["inline-react-svg", ["@emotion", { "cssPropOptimization": true }]]
-}
