--- packages/cna-template/template/components/Meta/index.ts
@@ -1,1 +1,0 @@
-export { Meta } from './Meta';
