--- packages/cna-template/template/jest.config.json
@@ -1,18 +1,0 @@
-{
-  "testURL": "http://localhost",
-  "moduleFileExtensions": ["js", "jsx", "ts", "tsx"],
-  "testPathIgnorePatterns": ["./.next/", "./node_modules/"],
-  "setupFilesAfterEnv": ["<rootDir>/setupTests.tsx"],
-  "snapshotSerializers": ["@emotion/jest/serializer"],
-  "coverageDirectory": "./__coverage__",
-  "coverageReporters": ["cobertura", "text-summary", "html"],
-  "collectCoverageFrom": [
-    "components/**/*.(js|jsx|ts|tsx)",
-    "pages/**/*.(js|jsx|ts|tsx)",
-    "services/**/*.(js|jsx|ts|tsx)",
-    "!components/**/index.(js|jsx|ts|tsx)",
-    "!**/*.d.ts",
-    "!/public/",
-    "!/node_modules/"
-  ]
-}
