+++ packages/cna-template/template/package.json
@@ -1,52 +1,24 @@
 {
-  "name": "next-app",
-  "version": "1.0.0",
-  "license": "ISC",
+  "name": "template-new",
+  "version": "0.1.0",
+  "private": true,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "analyze": "ANALYZE=true yarn build",
-    "test": "jest --watch",
-    "test:ci": "mkdir -p __reports__ && jest --ci --coverage --runInBand --json --outputFile=__reports__/jest-results.json",
-    "postinstall": "yarn foundry init -p lint -l TypeScript -e Browser Node; echo \"Remove the postinstall script from the package.json file after setting up Foundry.\""
+    "lint": "next lint"
   },
   "dependencies": {
-    "@emotion/css": "^11.1.3",
-    "@emotion/react": "^11.4.1",
-    "@emotion/server": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
-    "@next/bundle-analyzer": "^11.1.2",
-    "@sumup/circuit-ui": "^4.0.0",
-    "@sumup/collector": "^1.0.0",
-    "@sumup/design-tokens": "^3.1.0",
-    "@sumup/icons": "^2.0.0",
-    "@sumup/intl": "^1.0.0",
-    "next": "^11.1.2",
-    "next-transpile-modules": "^8.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "next": "12.1.6",
+    "react": "18.1.0",
+    "react-dom": "18.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/runtime": "^7.0.0",
-    "@emotion/babel-plugin": "^11.3.0",
-    "@emotion/jest": "^11.3.0",
-    "@sumup/foundry": "^4.0.2",
-    "@testing-library/jest-dom": "^5.0.0",
-    "@testing-library/react": "^11.0.0",
-    "@testing-library/react-hooks": "^3.0.0",
-    "@testing-library/user-event": "^12.0.0",
-    "@types/jest": "^26.0.0",
-    "@types/jest-axe": "^3.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/testing-library__jest-dom": "^5.0.0",
-    "babel-jest": "^26.0.0",
-    "babel-plugin-inline-react-svg": "^1.0.0",
-    "jest": "^26.0.0",
-    "jest-axe": "^4.0.0",
-    "react-test-renderer": "^17.0.0",
-    "typescript": "^4.4.3"
+    "@types/node": "17.0.31",
+    "@types/react": "18.0.8",
+    "@types/react-dom": "18.0.3",
+    "eslint": "8.14.0",
+    "eslint-config-next": "12.1.6",
+    "typescript": "4.6.4"
   }
 }
