+++ packages/cna-template/template/pages/api/hello.ts
