+++ packages/cna-template/template/pages/index.tsx
@@ -1,125 +1,72 @@
-import React from 'react';
-import { NextPage } from 'next';
-import { css } from '@emotion/react';
-import styled from '@emotion/styled';
-import {
-  Card,
-  Headline,
-  Body,
-  List,
-  Anchor,
-  cx,
-  spacing,
-} from '@sumup/circuit-ui';
+import type { NextPage } from 'next'
+import Head from 'next/head'
+import Image from 'next/image'
+import styles from '../styles/Home.module.css'
 
-import { Meta } from '../components/Meta';
-import { Logo } from '../components/Logo';
+const Home: NextPage = () => {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
 
-const Main = styled('main')(
-  ({ theme }) => css`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 100%;
-    max-width: 450px;
-    margin: 0 auto ${theme.spacings.mega};
-  `,
-);
+      <main className={styles.main}>
+        <h1 className={styles.title}>
+          Welcome to <a href="https://nextjs.org">Next.js!</a>
+        </h1>
 
-const centeredStyles = css`
-  text-align: center;
-`;
+        <p className={styles.description}>
+          Get started by editing{' '}
+          <code className={styles.code}>pages/index.tsx</code>
+        </p>
 
-const title = 'Welcome to SumUp Next.js';
+        <div className={styles.grid}>
+          <a href="https://nextjs.org/docs" className={styles.card}>
+            <h2>Documentation &rarr;</h2>
+            <p>Find in-depth information about Next.js features and API.</p>
+          </a>
 
-const Page: NextPage = () => (
-  <>
-    <Meta title={title} path="/" />
-    <Main>
-      <Logo />
-      <Card>
-        <Headline
-          size="one"
-          as="h1"
-          noMargin
-          css={cx(centeredStyles, spacing({ bottom: 'giga' }))}
+          <a href="https://nextjs.org/learn" className={styles.card}>
+            <h2>Learn &rarr;</h2>
+            <p>Learn about Next.js in an interactive course with quizzes!</p>
+          </a>
+
+          <a
+            href="https://github.com/vercel/next.js/tree/canary/examples"
+            className={styles.card}
+          >
+            <h2>Examples &rarr;</h2>
+            <p>Discover and deploy boilerplate example Next.js projects.</p>
+          </a>
+
+          <a
+            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+            className={styles.card}
+          >
+            <h2>Deploy &rarr;</h2>
+            <p>
+              Instantly deploy your Next.js site to a public URL with Vercel.
+            </p>
+          </a>
+        </div>
+      </main>
+
+      <footer className={styles.footer}>
+        <a
+          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+          target="_blank"
+          rel="noopener noreferrer"
         >
-          {title}
-        </Headline>
-        <Body noMargin css={spacing({ bottom: 'giga' })}>
-          This is a{' '}
-          <Anchor href="https://nextjs.org" target="_blank" noMargin>
-            Next.js
-          </Anchor>
-          -based starter project featuring some{' '}
-          <Anchor href="https://sumup.com" target="_blank" noMargin>
-            SumUp
-          </Anchor>
-          -specific customizations:
-        </Body>
-
-        <List size="one" noMargin css={spacing({ bottom: 'giga' })}>
-          <li>
-            <Anchor
-              href="https://github.com/sumup-oss/circuit-ui"
-              target="_blank"
-              noMargin
-            >
-              Circuit UI
-            </Anchor>{' '}
-            integration
-          </li>
-          <li>
-            <Anchor href="https://emotion.sh/" target="_blank" noMargin>
-              Emotion
-            </Anchor>{' '}
-            support (incl. babel plugin)
-          </li>
-          <li>
-            <Anchor href="https://lodash.com/" target="_blank" noMargin>
-              Lodash
-            </Anchor>{' '}
-            support (incl. babel plugin)
-          </li>
-          <li>
-            {"SumUp's default tooling provided by "}
-            <Anchor
-              href="https://github.com/sumup-oss/foundry"
-              target="_blank"
-              noMargin
-            >
-              Foundry
-            </Anchor>
-          </li>
-          <li>Basic SEO and performance optimizations</li>
-        </List>
-
-        <Body
-          size="one"
-          noMargin
-          css={cx(centeredStyles, spacing({ bottom: 'giga' }))}
-        >
-          Now go and build things!
-          <br />
-        </Body>
-        <Body
-          size="one"
-          noMargin
-          css={cx(
-            css`
-              ${centeredStyles};
-              transform: scale3d(1.5, 1.5, 1);
-            `,
-            spacing({ bottom: 'giga' }),
-          )}
-        >
-          <span role="img" aria-label="Emojis for building things">
-            🔨👩🏽‍💻👨🏼‍💻🚀
+          Powered by{' '}
+          <span className={styles.logo}>
+            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
           </span>
-        </Body>
-      </Card>
-    </Main>
-  </>
-);
+        </a>
+      </footer>
+    </div>
+  )
+}
 
-export default Page;
+export default Home
