+++ packages/cna-template/template/public/favicon.ico
