+++ packages/cna-template/template/styles/globals.css
