--- packages/cna-template/template/types/media.d.ts
@@ -1,5 +1,0 @@
-declare module '*.png';
-declare module '*.jpg';
-declare module '*.jpeg';
-declare module '*.svg';
-declare module '*.gif';
